+++ tests/test_visitors/test_ast/test_keywords/test_loops/test_lambda_in_loop.py
@@ -1,0 +1,51 @@
+# -*- coding: utf-8 -*-
+
+import pytest
+
+from wemake_python_styleguide.visitors.ast.keywords import (
+    LambdaInsideLoopViolation,
+    WrongLoopVisitor,
+)
+
+lambda_inside_for_loop = """
+for index in range(10):
+    print(lambda: index)
+"""
+
+nested_lambda_inside_for_loop = """
+for index in range(10):
+    if some:
+        print(lambda: index)
+"""
+
+lambda_inside_while_loop = """
+while True:
+    print(lambda: index)
+"""
+
+nested_lambda_inside_while_loop = """
+while True:
+    if some:
+        print(lambda: index)
+"""
+
+
+@pytest.mark.parametrize('code', [
+    lambda_inside_for_loop,
+    nested_lambda_inside_for_loop,
+    lambda_inside_while_loop,
+    nested_lambda_inside_while_loop,
+])
+def test_lambda_body(
+    assert_errors,
+    parse_ast_tree,
+    code,
+    default_options,
+):
+    """Ensures that lambda can not be inside a loop's body."""
+    tree = parse_ast_tree(code)
+
+    visitor = WrongLoopVisitor(default_options, tree=tree)
+    visitor.run()
+
+    assert_errors(visitor, [LambdaInsideLoopViolation])